chore: 更新编译选项和库依赖
添加 `-Wno-unused-parameter` 编译选项以忽略未使用的参数警告
更新 MariaDB 库依赖为 `unofficial-libmariadb`
修改 `reader_list` 初始化方式以使用统一的初始化语法
将 `escape` 函数返回值从 `std::move(ret)` 改为 `ret`
修正 `compressionZipFile` 函数中 `pos` 类型为 `size_t` 以匹配 `find_last_of` 返回值类型

+++ src/utils/utils.cpp
@@ -15,7 +15,7 @@ Utils::Utils() {}
 Utils::~Utils() {}
 
 bool Utils::compressionZipFile(const std::string &file_path, bool remove) {
-    int pos = file_path.find_last_of("/");
+    size_t pos = file_path.find_last_of("/");
     if (pos == std::string::npos) return false;
     std::string save_path = file_path.substr(0, pos);
     return compressionZipFile(file_path, save_path, remove);
